+++ packages/google_sign_in/test/google_sign_in_test.dart
@@ -48,8 +48,10 @@ void main() {
       expect(
           log,
           equals(<MethodCall>[
-            new MethodCall('init',
-                <String, dynamic>{'scopes': <String>[], 'hostedDomain': null}),
+            new MethodCall('init', <String, dynamic>{
+              'scopes': <String>[],
+              'hostedDomain': null,
+            }),
             const MethodCall('signInSilently'),
           ]));
     });
@@ -60,8 +62,10 @@ void main() {
       expect(
           log,
           equals(<MethodCall>[
-            new MethodCall('init',
-                <String, dynamic>{'scopes': <String>[], 'hostedDomain': null}),
+            new MethodCall('init', <String, dynamic>{
+              'scopes': <String>[],
+              'hostedDomain': null,
+            }),
             const MethodCall('signIn'),
           ]));
     });
@@ -72,8 +76,10 @@ void main() {
       expect(
           log,
           equals(<MethodCall>[
-            new MethodCall('init',
-                <String, dynamic>{'scopes': <String>[], 'hostedDomain': null}),
+            new MethodCall('init', <String, dynamic>{
+              'scopes': <String>[],
+              'hostedDomain': null,
+            }),
             const MethodCall('signOut'),
           ]));
     });
@@ -84,8 +90,10 @@ void main() {
       expect(
           log,
           equals(<MethodCall>[
-            new MethodCall('init',
-                <String, dynamic>{'scopes': <String>[], 'hostedDomain': null}),
+            new MethodCall('init', <String, dynamic>{
+              'scopes': <String>[],
+              'hostedDomain': null,
+            }),
             const MethodCall('disconnect'),
           ]));
     });
@@ -97,8 +105,10 @@ void main() {
       expect(
           log,
           equals(<MethodCall>[
-            new MethodCall('init',
-                <String, dynamic>{'scopes': <String>[], 'hostedDomain': null}),
+            new MethodCall('init', <String, dynamic>{
+              'scopes': <String>[],
+              'hostedDomain': null,
+            }),
             const MethodCall('disconnect'),
           ]));
     });
@@ -120,8 +130,10 @@ void main() {
       expect(
           log,
           equals(<MethodCall>[
-            new MethodCall('init',
-                <String, dynamic>{'scopes': <String>[], 'hostedDomain': null}),
+            new MethodCall('init', <String, dynamic>{
+              'scopes': <String>[],
+              'hostedDomain': null,
+            }),
             const MethodCall('signInSilently'),
           ]));
     });
@@ -134,8 +146,10 @@ void main() {
       expect(
           log,
           equals(<MethodCall>[
-            new MethodCall('init',
-                <String, dynamic>{'scopes': <String>[], 'hostedDomain': null}),
+            new MethodCall('init', <String, dynamic>{
+              'scopes': <String>[],
+              'hostedDomain': null,
+            }),
             const MethodCall('signInSilently'),
             const MethodCall('signIn'),
           ]));
@@ -152,8 +166,10 @@ void main() {
       expect(
           log,
           equals(<MethodCall>[
-            new MethodCall('init',
-                <String, dynamic>{'scopes': <String>[], 'hostedDomain': null}),
+            new MethodCall('init', <String, dynamic>{
+              'scopes': <String>[],
+              'hostedDomain': null,
+            }),
             const MethodCall('signInSilently'),
           ]));
       expect(users.first, users.last, reason: 'Must return the same user');
@@ -179,8 +195,10 @@ void main() {
       expect(
           log,
           equals(<MethodCall>[
-            new MethodCall('init',
-                <String, dynamic>{'scopes': <String>[], 'hostedDomain': null}),
+            new MethodCall('init', <String, dynamic>{
+              'scopes': <String>[],
+              'hostedDomain': null,
+            }),
             const MethodCall('signOut'),
             const MethodCall('signOut'),
             const MethodCall('disconnect'),
@@ -200,8 +218,10 @@ void main() {
       expect(
           log,
           equals(<MethodCall>[
-            new MethodCall('init',
-                <String, dynamic>{'scopes': <String>[], 'hostedDomain': null}),
+            new MethodCall('init', <String, dynamic>{
+              'scopes': <String>[],
+              'hostedDomain': null,
+            }),
             const MethodCall('signInSilently'),
             const MethodCall('signOut'),
             const MethodCall('signIn'),