+++ packages/firebase_analytics/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.3
+
+* Updated to Firebase SDK Version 11.0.1
+
 ## 0.0.2
 
 * Bumped buildToolsVersion to 25.0.3