+++ packages/google_sign_in/CHANGELOG.md
@@ -1,3 +1,23 @@
+## 0.2.0
+
+* Updated dependencies
+* **Breaking Change**: You need to add a maven section with the "https://maven.google.com" endpoint to the repository section of your `android/build.gradle`. For example:
+```gradle
+allprojects {
+    repositories {
+        jcenter()
+        maven {                              // NEW
+            url "https://maven.google.com"   // NEW
+        }                                    // NEW
+    }
+}
+```
+
+## 0.1.0
+
+* Update to use `GoogleSignIn` CocoaPod
+
+
 ## 0.0.6
 
 * Fix crash on iOS when signing in caused by nil uiDelegate